pyproject.toml
[project]
name = "nsfw-guard"
version = "0.1.0"
description = "Simple nude image AI detectoк"
authors = [
    {name = "dmvaled",email = "dmvaled@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.14,<0.116.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "requests (>=2.32.4,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.2"
pre-commit = "^4.0.1"


[tool.poetry.group.test.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[tool.ruff]
line-length = 88
target-version = "py312"
fix = true
unsafe-fixes = true

[tool.ruff.lint]
extend-select = ["E", "F", "W", "I"]

[tool.ruff.lint.isort]
force-single-line = true
combine-as-imports = false
force-sort-within-sections = true
lines-after-imports = 2

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
